+++ CMakeLists.txt
@@ -179,6 +179,11 @@ google_binary(cartographer_compute_relations_metrics
     cartographer/ground_truth/compute_relations_metrics_main.cc
 )
 
+google_binary(cartographer_migrate_serialized_format
+  SRCS
+    cartographer/tools/migrate_stream_format_main.cc
+)
+
 if(${BUILD_GRPC})
   google_binary(cartographer_grpc_server
     SRCS
